All tests and benchmarks migrated from tests/tnt to tarantool_test.go
