--- tests/tnt/doc.go
@@ -1,1 +1,0 @@
-package tnt
